// src/app/api/model/transactionRead.ts
/**
 * FinanzApp API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransactionTypeEnum } from './transactionTypeEnum';

export interface TransactionRead {
  id: string;
  account_id: string;
  to_account_id: string | null;
  category_id: string | null;
  transfer_group_id: string | null;
  payment_plan_id: string | null;
  amount: number;
  type: TransactionTypeEnum;
  date: string;
  notes: string | null;
  created_by: string | null;
  updated_by: string | null;
  created_at: string;
  updated_at: string;
}
export namespace TransactionRead {}