// src/app/api/model/updateTransactionRequest.ts
/**
 * FinanzApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransactionTypeEnum } from './transactionTypeEnum';


export interface UpdateTransactionRequest { 
    amount?: number | null;
    type?: TransactionTypeEnum | null;
    category_id?: string | null;
    date?: string | null;
    notes?: string | null;
}
export namespace UpdateTransactionRequest {
}